Completely remove aria button, and use "normal" button instead.
